# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Mycorrhizal CRM API
  description: >
    REST API for Mycorrhizal's contact model. As of WP-71 (P2,
    docs/fork-plan/50-integration-and-rebrand.md), the contact endpoints speak
    the neutral RFC 9553-shaped model (contactmodel.Record/Card) rather than
    the legacy flat Contact DTO. GET /contacts (list) returns a slim
    ContactSummary per item; GET/POST/PUT /contacts/{id} exchange the full
    Card/CRM shape (ContactRecordResponse / ContactRecordInput) — a
    symmetric read/write contract.


    T8 (docs/fork-plan/tickets/16-T8-openapi.md) extended this spec to cover
    the whole REST surface: every route registered by
    backend/routes/routes.go under /api/v1 plus the unversioned /health. The
    drift test (backend/openapi_test.go) enumerates the live Gin router and
    fails on any registered route with no spec entry, so this file cannot go
    stale again.

    Coverage notes:

    - **Auth:** every path under /api/v1 except the public auth group
      (register, login, logout, check-password-strength, password-reset,
      auth/oidc/config) requires a session cookie (`auth_token`, set by login
      and OIDC callback) or an `Authorization: Bearer <token>` header. API
      tokens (mycorrhizal_…) are accepted on the general REST path only when
      their scope is `full`; a `carddav`-scoped token is rejected with 403.
      API tokens cannot access the admin paths at all.
...
      `/auth/oidc/config` is always present.
...
      them.
...
servers:
...
security:
  - sessionCookie: []
  - bearerAuth: []
paths:
...
  /life-events:
    get:
      summary: List life events
      operationId: listLifeEvents
      parameters:
        - name: cursor
          in: query
          description: Opaque resume token from the previous page's next_cursor. Ignored when since is present.
          schema: { type: string }
        - name: since
          in: query
          description: >
            Change-feed cursor: return every life event changed after this
            position — including soft-deleted ones (deleted:true) — ordered
            forward, ignoring entity_id. A cursor older than
            DELETED_RETENTION_DAYS returns 410 Gone.
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, default: 25 }
        - name: order
          in: query
          description: Direction for the (updated_at, id) cursor order. since forces asc.
          schema: { type: string, enum: [asc, desc], default: desc }
        - name: entity_id
          in: query
          description: Filter to life events whose subject contact is this Contact.VCardUID.
          schema: { type: string, format: uuid }
      responses:
        "401":
          description: Not authenticated (missing or invalid session cookie or full-scope bearer token).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "410":
          description: The ?since= cursor is older than DELETED_RETENTION_DAYS — full resync required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "200":
          description: A page of life events.
          content:
            application/json:
              schema:
                type: object
                properties:
                  life_events:
                    type: array
                    items:
                      $ref: "#/components/schemas/LifeEvent"
                  next_cursor:
                    type: string
                    description: Resume token for the next page; empty when there are no more rows.
                  limit: { type: integer }
                  sync:
                    $ref: "#/components/schemas/SyncInfo"
    post:
      summary: Create a life event
      operationId: createLifeEvent
      description: When remind is true and the date has a month/day, a yearly reminder is materialized.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LifeEventInput"
      responses:
        "201":
          description: Life event created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: "Life event created successfully" }
                  life_event:
                    $ref: "#/components/schemas/LifeEvent"
        "400":
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: The subject contact was not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /life-events/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: Get a life event
      operationId: getLifeEvent
      responses:
        "200":
          description: The life event (raw).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LifeEvent"
        "404":
          description: Life event not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      summary: Update a life event (full replace)
      operationId: updateLifeEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LifeEventInput"
      responses:
        "200":
          description: The updated life event (raw).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LifeEvent"
        "400":
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Life event (or subject contact) not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete a life event (soft delete; its materialized reminder hard-deletes)
      operationId: deleteLifeEvent
      responses:
        "200":
          description: Deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Life event not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # --- Preferences ---
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: auth_token
      description: The httpOnly session cookie set by POST /login and the OIDC callback.
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        A JWT or a `mycorrhizal_…` API token. On the general REST path a
        carddav-scoped token is rejected with 403 (only `full`-scope tokens
        work); API tokens are rejected on every /admin path entirely.
  schemas:
...
    Error:
      type: object
      description: The standard error envelope produced by apperrors.AbortWithError.
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: object
              additionalProperties: true
        request_id: { type: string }
        timestamp: { type: string, format: date-time }
    SyncInfo:
      type: object
      description: >
        T17 sync-mode contract carried by every cursor-paginated list
        response. `mode` is how THIS collection syncs; the two arrays are the
        static map of every syncable collection, so a client can plan a full
        sync without hardcoding endpoint names. Incremental collections are
        soft-deleted, support ?since= change feeds, and drop `total`;
        full_resync collections are hard-deleted, are re-pulled wholesale, and
        keep `total`.
      properties:
        mode: { type: string, enum: [incremental, full_resync] }
        incremental:
          type: array
          items: { type: string }
          description: "Collections that sync incrementally via ?since=: contacts, notes, activities, life_events, preferences."
        full_resync:
          type: array
          items: { type: string }
          description: "Collections that require a full re-pull: circles, households, tags, relationship_edges, field_definitions."
    MessageResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }
      additionalProperties: true
...
    LifeEvent:
      type: object
      description: A permanent fact about an entity's life. remind=false is omitted on the wire (omitempty).
      properties:
        id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        entity_id: { type: string, format: uuid }
        type: { type: string }
        date:
          $ref: "#/components/schemas/PartialDate"
        description: { type: string, maxLength: 2000 }
        source: { type: string, enum: [user, imported, ai-suggested] }
        related_entity_ids:
          type: array
          items: { type: string, format: uuid }
        remind: { type: boolean }
        deleted:
          type: boolean
          description: "Change-feed tombstone marker (T17): true only via ?since= for a soft-deleted life event."
          default: false
    LifeEventInput:
      type: object
      required: [entity_id]
      properties:
        entity_id: { type: string, format: uuid }
        type: { type: string }
        date:
          $ref: "#/components/schemas/PartialDate"
        description: { type: string, maxLength: 2000 }
        source: { type: string, enum: [user, imported, ai-suggested] }
        related_entity_ids:
          type: array
          items: { type: string, format: uuid }
        remind: { type: boolean }

    # --- Preferences ---
...
    PartialDate:
      type: object
      properties:
        year: { type: integer, nullable: true }
        month: { type: integer, nullable: true }
        day: { type: integer, nullable: true }
        calendarScale: { type: string }
...